added class method for order to get cart
